Fixed more CI warnings

+++ src/config/base.cpp
@@ -204,7 +204,7 @@ std::vector<std::string> ConfigBase::_merge(
     //   might be our current config, or might be one single one of the new incoming messages).
     // - confs that failed to parse (we can't understand them, so leave them behind as they may be
     //   some future message).
-    int superconf = new_conf->unmerged_index();  // -1 if we had to merge
+    size_t superconf = new_conf->unmerged_index();  // -1 if we had to merge
     for (size_t i = 0; i < all_hashes.size(); i++) {
         if (i != superconf && !bad_confs.count(i) && !all_hashes[i].empty())
             _old_hashes.emplace(all_hashes[i]);
